pylintrc: re-enable too-many-lines, part 3
To get under the 1000 line limit, move some ycheck
scenario helpers to tests/unit/utils.py.
